+++ src/common/db/migrations/1715941788971-create-users-voices.ts
@@ -7,10 +7,10 @@ export class CreateUsersVoices1715941788971 implements MigrationInterface {
                 name: 'users_voices',
                 columns: [
                     { name: 'id', type: 'int', isPrimary: true, isGenerated: true, generationStrategy: 'increment' },
-                    { name: 'duration', type: 'int', isNullable: false },
+                    { name: 'duration', type: 'bigint', isNullable: false },
                     { name: 'file_id', type: 'varchar', isNullable: false, isUnique: true },
                     { name: 'file_unique_id', type: 'varchar', isNullable: false, isUnique: true },
-                    { name: 'file_size', type: 'int', isNullable: false },
+                    { name: 'file_size', type: 'bigint', isNullable: false },
                     { name: 'user_id', type: 'int', isNullable: false },
                     { name: 'chat_id', type: 'int', isNullable: false },
                     { name: 'created_at', type: 'timestamptz', isNullable: false, default: 'now()' },
